+++ src/App.tsx
@@ -2,12 +2,12 @@ import * as React from 'react';
 import './i18n';
 import AppAppBar from './app-appbar';
 import SVGs from './svgs';
-// import SVGs from './svgs';
 import Panels from './panels';
 import { getParams } from './utils';
 import { getBranches, useTpo, getRoutes } from './methods';
 import { CanvasContext, ParamContext, paramReducer } from './context';
 import { createMuiTheme, ThemeProvider, useMediaQuery, LinearProgress } from '@material-ui/core';
+import { ProvidedCanvas } from './types';
 
 const darkTheme = createMuiTheme({
     palette: {
@@ -53,24 +53,27 @@ const lightTheme = createMuiTheme({
     },
 });
 
-export default function App() {
+export default function App(props: { canvas: ProvidedCanvas[] }) {
     const prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)');
     const theme = prefersDarkMode ? darkTheme : lightTheme;
 
-    const canvasAvailable = ((style): ('destination' | 'runin' | 'railmap')[] => {
-        switch (style) {
-            case 'mtr':
-                return ['destination', 'railmap'];
-            case 'gzmtr':
-                return ['runin', 'railmap'];
-            case 'shmetro':
-                return ['destination', 'runin', 'railmap'];
-            default:
-                return [];
-        }
-    })(window.urlParams.get('style'));
-    const [canvasToShown, setCanvasToShown] = React.useState<'all' | 'destination' | 'runin' | 'railmap'>('all');
+    const [canvasToShown, setCanvasToShown] = React.useState<'all' | ProvidedCanvas>('all');
 
+    return (
+        <>
+            <ThemeProvider theme={theme}>
+                <CanvasContext.Provider value={{ canvasAvailable: props.canvas, canvasToShown, setCanvasToShown }}>
+                    <React.Suspense fallback={<LinearProgress />}>
+                        <AppAppBar />
+                    </React.Suspense>
+                    <AppBody />
+                </CanvasContext.Provider>
+            </ThemeProvider>
+        </>
+    );
+}
+
+const AppBody = () => {
     const [param, dispatch] = React.useReducer(paramReducer, getParams());
     const paramString = JSON.stringify(param);
     React.useEffect(() => localStorage.setItem('rmgParam', paramString), [paramString]);
@@ -94,27 +97,20 @@ export default function App() {
     const handleUpdate = (key, data) => dispatch({ type: 'ANY', key, data });
 
     return (
-        <>
-            <ThemeProvider theme={theme}>
-                <CanvasContext.Provider value={{ canvasAvailable, canvasToShown, setCanvasToShown }}>
-                    <React.Suspense fallback={<LinearProgress />}>
-                        <AppAppBar />
-                    </React.Suspense>
-                    <ParamContext.Provider
-                        value={{
-                            param,
-                            dispatch,
-                            branches,
-                            routes,
-                            deps,
-                        }}
-                    >
-                        <SVGs />
-                        <Panels param={param} paramUpdate={handleUpdate} tpo={tpo} />
-                    </ParamContext.Provider>
-                </CanvasContext.Provider>
-                <canvas style={{ display: 'none' }} />
-            </ThemeProvider>
-        </>
+        <div style={{ flex: 1, overflow: 'overlay' }}>
+            <ParamContext.Provider
+                value={{
+                    param,
+                    dispatch,
+                    branches,
+                    routes,
+                    deps,
+                }}
+            >
+                <SVGs />
+                <Panels param={param} paramUpdate={handleUpdate} tpo={tpo} />
+            </ParamContext.Provider>
+            <canvas style={{ display: 'none' }} />
+        </div>
     );
-}
+};