+++ src/index3.tsx
@@ -1,14 +1,13 @@
-import * as $ from 'jquery';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import App from './App';
 import { updateParam } from './utils';
+import { ProvidedCanvas } from './types';
 
 declare global {
     interface Window {
         urlParams?: URLSearchParams;
         gtag: any;
-        // test3: any;
     }
 }
 
@@ -22,25 +21,43 @@ switch (window.urlParams.get('style')) {
 }
 history.pushState({ url: window.location.href }, null, '?' + window.urlParams.toString());
 
+const canvasAvailable = ((style): ProvidedCanvas[] => {
+    switch (style) {
+        case 'mtr':
+            return ['destination', 'railmap'];
+        case 'gzmtr':
+            return ['runin', 'railmap'];
+        case 'shmetro':
+            return ['destination', 'runin', 'railmap'];
+        default:
+            return [];
+    }
+})(window.urlParams.get('style'));
+
 // load stylesheets on demand
-$('head').append(
-    ...['share', 'destination', 'runin', 'railmap'].map(tag => {
-        return $('<link>', {
-            rel: 'stylesheet',
-            href: `styles/${tag}_${window.urlParams.get('style')}.css`,
-            id: `css_${tag}`,
-        });
+document.head.append(
+    ...['share', ...canvasAvailable].map(tag => {
+        let link = document.createElement('link');
+        link.rel = 'stylesheet';
+        link.href = `styles/${tag}_${window.urlParams.get('style')}.css`;
+        link.id = `css_${tag}`;
+        return link;
     })
 );
+
+const renderApp = () => {
+    ReactDOM.render(<App canvas={canvasAvailable} />, document.querySelectorAll('div#app')[0]);
+};
 
 if (localStorage.rmgParam) {
     try {
         updateParam();
     } catch (err) {
-        alert(err + 'Something error! Please start from a new canvas. ');
+        alert(err + 'Something error! Please refresh to start from a new canvas. ');
         console.error(err);
+        localStorage.removeItem('rmgParam');
     }
-    ReactDOM.render(<App />, $('#app')[0]);
+    renderApp();
 } else {
     fetch('templates/blank.json')
         .then(response => response.json())
@@ -48,5 +65,5 @@ if (localStorage.rmgParam) {
             localStorage.setItem('rmgParam', JSON.stringify(data));
             updateParam();
         })
-        .then(() => ReactDOM.render(<App />, $('#app')[0]));
+        .then(() => renderApp());
 }