+++ src/panels/save/export-diag/preview-diag.tsx
@@ -13,85 +13,87 @@ interface Props {
 export default (props: Props) => {
     const { t } = useTranslation();
 
-    const [svgEl, setSvgEl] = React.useState(($('<svg>')[0] as Element) as SVGSVGElement);
+    const [svgEl, setSvgEl] = React.useState((document.createElement('svg') as Element) as SVGSVGElement);
     const [isLoaded, setIsLoaded] = React.useState(false);
+
+    const contentEl = React.useRef<HTMLDivElement>();
 
     React.useEffect(() => {
         if (props.canvas === '') {
-            setSvgEl(($('<svg>')[0] as Element) as SVGSVGElement);
+            setSvgEl((document.createElement('svg') as Element) as SVGSVGElement);
             setIsLoaded(false);
             return;
         }
-        let [thisSVGWidth, thisSVGHeight] = [
-            $(`svg#${props.canvas}`)
-                .css('--rmg-svg-width')
-                .match(/\d+/g),
-            $(`svg#${props.canvas}`)
-                .css('--rmg-svg-height')
-                .match(/\d+/g),
-        ].map(Number);
-        let MAX_WIDTH = Math.min($(window).width(), 1412) - 64 - 24 * 2;
-        let MAX_HEIGHT = $(window).height() - 64 - 64 - 52 - 8 * 2;
+        let [thisSVGWidth, thisSVGHeight] = ['--rmg-svg-width', '--rmg-svg-height']
+            .map(
+                key =>
+                    (document.querySelector(`svg#${props.canvas}`) as SVGSVGElement).style
+                        .getPropertyValue(key)
+                        .match(/\d+/g)[0]
+            )
+            .map(Number);
+
+        let MAX_WIDTH = Math.min(window.innerWidth, 1412) - 64 - 24 * 2;
+        let MAX_HEIGHT = window.innerHeight - 64 - 64 - 52 - 8 * 2;
         let scaleFactor = Math.min(MAX_WIDTH / thisSVGWidth, MAX_HEIGHT / thisSVGHeight);
 
-        let el = $(`svg#${props.canvas}`)
-            .clone()
-            .attr({
-                viewBox: `0 0 ${thisSVGWidth} ${thisSVGHeight}`,
-                width: thisSVGWidth * scaleFactor,
-                height: thisSVGHeight * scaleFactor,
-            })
-            .css({
-                all: 'initial',
-            });
+        let elem = document.querySelector(`svg#${props.canvas}`).cloneNode(true) as SVGSVGElement;
+        elem.viewBox.baseVal.width = thisSVGWidth;
+        elem.viewBox.baseVal.height = thisSVGHeight;
+        elem.setAttribute('width', (thisSVGWidth * scaleFactor).toString());
+        elem.setAttribute('height', (thisSVGHeight * scaleFactor).toString());
+        elem.style.setProperty('all', 'initial');
 
         let cssTxt = ['share', props.canvas].map(tag => {
-            return Array.from((($(`link#css_${tag}`)[0] as HTMLLinkElement).sheet as CSSStyleSheet).cssRules)
+            return Array.from(
+                (Array.from(document.querySelectorAll('link')).filter(l => l.id === 'css_' + tag)[0]
+                    ?.sheet as CSSStyleSheet).cssRules
+            )
                 .map(rule => rule.cssText)
                 .join(' ');
         });
-        el.prepend(...cssTxt.map(txt => $('<style>').text(txt)))
-            .prepend(document.querySelector('style#global').outerHTML)
-            .find('[style="display: none;"]')
-            .remove();
+
+        cssTxt.forEach(txt => {
+            let s = document.createElement('style');
+            s.innerText = txt;
+            elem.prepend(s);
+        });
 
         if (window.urlParams.get('style') === 'mtr') {
             import(/* webpackChunkName: "panelPreviewMTR" */ './mtr-helper').then(({ getBase64FontFace }) => {
-                getBase64FontFace((el[0] as Element) as SVGSVGElement).then(async response => {
+                getBase64FontFace(elem).then(async response => {
                     let uris = await Promise.all(response);
-                    el.prepend($('<style>').text(uris.join(' ')));
+                    let s = document.createElement('style');
+                    s.innerText = uris.join(' ');
+                    elem.prepend(s);
 
-                    setSvgEl((el[0] as Element) as SVGSVGElement);
-
-                    document.fonts.ready.then(() => {
-                        setIsLoaded(true);
-                    });
+                    setSvgEl(elem);
+                    document.fonts.ready.then(() => setIsLoaded(true));
                 });
             });
         } else {
-            setSvgEl((el[0] as Element) as SVGSVGElement);
+            setSvgEl(elem);
             setIsLoaded(true);
         }
     }, [props.canvas]);
 
     const handleClose = (action: string) => () => {
+        let svgEl = contentEl.current.querySelector('svg');
         if (action === 'png') {
-            test($('#preview_diag svg'));
+            test(svgEl);
         } else if (action === 'svg') {
-            let svgContent = $('#preview_diag svg');
-
             var link = document.createElement('a');
-            link.href = 'data:image/svg+xml;base64,' + btoa(unescape(encodeURIComponent(svgContent[0].outerHTML)));
-            link.download = 'rmg_export.svg';
+            link.href = 'data:image/svg+xml;base64,' + btoa(unescape(encodeURIComponent(svgEl.outerHTML)));
+            link.download = 'rmg.' + new Date().toISOString() + '.svg';
             link.click();
         }
         props.onClose('close');
     };
 
     return (
-        <Dialog onClose={handleClose('close')} open={props.open} maxWidth={false} id="preview_diag">
+        <Dialog onClose={handleClose('close')} open={props.open} maxWidth={false}>
             <DialogTitle>{t('file.preview.title')}</DialogTitle>
-            <DialogContent dangerouslySetInnerHTML={{ __html: svgEl.outerHTML }} />
+            <DialogContent dangerouslySetInnerHTML={{ __html: svgEl.outerHTML }} ref={contentEl} />
             <DialogActions>
                 <Button onClick={handleClose('close')} color="primary" autoFocus>
                     {t('dialog.cancel')}