+++ src/types.ts
@@ -1,23 +1,30 @@
-export enum DirectionLong {left, right};
-export enum NeighbourPl {parents, children};
+export type ProvidedCanvas = 'destination' | 'runin' | 'railmap';
+
+export enum DirectionLong {
+    left,
+    right,
+}
+export enum NeighbourPl {
+    parents,
+    children,
+}
 /**
- * Information of branch on both side. 
+ * Information of branch on both side.
  * @index 0 - branch type (through, nonthrough or empty)
  * @index 1 - ID of the first station of the branch (should also be one of the station's parents/children) or empty
  */
 export interface BranchInfo {
-    left: ['through' | 'nonthrough', string] | [], 
-    right: ['through' | 'nonthrough', string] | []
+    left: ['through' | 'nonthrough', string] | [];
+    right: ['through' | 'nonthrough', string] | [];
 }
 /**
- * Array of name `string`s. The first element is in Chinese characters and the second element is in Latin characters. 
+ * Array of name `string`s. The first element is in Chinese characters and the second element is in Latin characters.
  */
 export type Name = [string, string];
 
-
 export interface StationInfo {
     /**
-     * Station name in two languages. 
+     * Station name in two languages.
      */
     name: Name;
     /**
@@ -25,60 +32,55 @@ export interface StationInfo {
      */
     num?: string;
     /**
-     * Dictionary of the information of branch on the station's both side. 
+     * Dictionary of the information of branch on the station's both side.
      */
     branch?: BranchInfo;
     /**
-     * Array of parents' IDs. 
+     * Array of parents' IDs.
      */
     parents: string[];
     /**
-     * Array of children's IDs. 
+     * Array of children's IDs.
      */
     children: string[];
     /**
-     * Detail of interchanges. 
+     * Detail of interchanges.
      */
     transfer?: StationTransfer;
     /**
-     * Array of services at this station. 
+     * Array of services at this station.
      */
     services: ('local' | 'express')[];
     /**
-     * Detail of interchanges (legacy). 
+     * Detail of interchanges (legacy).
      */
     interchange?: any;
     change_type?: string;
     usage?: 'airport' | 'disney' | 'hsr' | '';
     facility?: 'airport' | 'disney' | 'hsr' | '';
-};
+}
 
-export enum IntInfoTag {
-    city, line, colour, fg, nameZH, nameEN
-};
-export type InterchangeInfo = {
-    [T in IntInfoTag]: string;
-};
+export type InterchangeInfo = [string, string, string, '#000' | '#fff', string, string];
 
 export interface StationTransfer {
     /**
-     * Interchange type of station. 
+     * Interchange type of station.
      */
     type: 'none' | 'int2' | 'int3' | 'osi11' | 'osi12' | 'osi13' | 'osi21' | 'osi22' | 'osi31';
     /**
-     * Direction of text/tick of interchanges. 
+     * Direction of text/tick of interchanges.
      */
     tick_direc: 'r' | 'l';
     /**
-     * Flag of paid area within out-of-station interchange. 
+     * Flag of paid area within out-of-station interchange.
      */
     paid_area: boolean;
     /**
-     * Array of name (in two languages) of all out-of-station interchange stations. 
+     * Array of name (in two languages) of all out-of-station interchange stations.
      */
     osi_names: Name[];
     /**
-     * Array of arrays of interchange info. 
+     * Array of arrays of interchange info.
      * @index 0 - array of within-station interchange info
      * @index remaining - arrays of out-of-station interchange info (from the nearest to the furthest station)
      */
@@ -86,7 +88,7 @@ export interface StationTransfer {
 }
 
 /**
- * Dictionary of configuration parameters for RMG, stored in `localStorage` as string. 
+ * Dictionary of configuration parameters for RMG, stored in `localStorage` as string.
  */
 export interface RMGParam {
     /**
@@ -102,11 +104,11 @@ export interface RMGParam {
      */
     svg_height: number;
     /**
-     * Train direction. 
+     * Train direction.
      */
     direction: 'l' | 'r';
     /**
-     * Colour theme of line, derived from `LineEntry`
+     * Colour theme of line, derived from `LineEntry`.
      * @index 0 - city id
      * @index 1 - line id
      * @index 2 - background colour (in HEX)
@@ -114,11 +116,15 @@ export interface RMGParam {
      */
     theme: [string, string, string, '#fff' | '#000'];
     /**
-     * ID of current station. 
+     * Name of line.
+     */
+    line_name: Name;
+    /**
+     * ID of current station.
      */
     current_stn_idx: string;
     /**
-     * Key-value pairs of the information of each station. 
+     * Key-value pairs of the information of each station.
      */
     stn_list: {
         [stnId: string]: StationInfo;
@@ -136,33 +142,33 @@ export interface RMGParam {
      */
     char_form?: 'trad' | 'cn' | 'tw' | 'jp';
     [propName: string]: any;
-};
+}
 
 export interface LineEntry {
     /**
-     * ID of line. 
+     * ID of line.
      */
     id: string;
     /**
-     * Key-value pairs of multi-lingual names of the line. 
+     * Key-value pairs of multi-lingual names of the line.
      */
     name: {
         en: string;
         [lang: string]: string;
     };
     /**
-     * Background colour (in #HEX). 
+     * Background colour (in #HEX).
      */
     colour: string;
     /**
-     * Foreground colour. Mandatory field if foreground colour is black. 
+     * Foreground colour. Mandatory field if foreground colour is black.
      */
     fg?: '#000' | '#fff';
-};
+}
 
 export interface CityEntry {
     /**
-     * ID of city. 
+     * ID of city.
      */
     id: string;
     /**
@@ -170,10 +176,10 @@ export interface CityEntry {
      */
     country: string;
     /**
-     * Key-value pairs of multi-lingual names of the city. 
+     * Key-value pairs of multi-lingual names of the city.
      */
     name: {
         en: string;
         [lang: string]: string;
-    }
-};
+    };
+}